Extract the `buckets` module
